[refactor] move ethcall function to UniswapService
