+++ src/main/java/com/oneswap/service/impl/UniswapServiceImpl.java
@@ -2,6 +2,7 @@ package com.oneswap.service.impl;
 
 import com.oneswap.model.Liquidity;
 import com.oneswap.service.UniswapService;
+import lombok.extern.log4j.Log4j2;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.beans.factory.annotation.Value;
@@ -9,6 +10,7 @@ import org.springframework.stereotype.Service;
 import org.web3j.abi.FunctionEncoder;
 import org.web3j.abi.FunctionReturnDecoder;
 import org.web3j.abi.TypeReference;
+import org.web3j.abi.datatypes.Address;
 import org.web3j.abi.datatypes.Function;
 import org.web3j.abi.datatypes.Type;
 import org.web3j.abi.datatypes.generated.Uint256;
@@ -24,6 +26,7 @@ import java.util.Collections;
 import java.util.List;
 
 @Service
+@Log4j2
 public class UniswapServiceImpl implements UniswapService {
 
     @Value("${blockchain}")
@@ -93,5 +96,43 @@ public class UniswapServiceImpl implements UniswapService {
 
         List<Type> decodedResponse = FunctionReturnDecoder.decode(response.getValue(), function.getOutputParameters());
         return (String) decodedResponse.get(0).getValue();
+    }
+
+    public String getToken0(String contractAddress) {
+        Function function = new Function("token0", Arrays.asList(), Arrays.asList(new TypeReference<Address>() {
+        }));
+        String token0 = null;
+        try {
+            token0 = callContractFunction(function, contractAddress);
+        } catch (Exception e) {
+            log.warn(e.getMessage());
+        }
+        return token0;
+    }
+
+    public String getToken1(String contractAddress) {
+        Function function = new Function("token1", Arrays.asList(), Arrays.asList(new TypeReference<Address>() {
+        }));
+        String token1 = null;
+        try {
+            token1 = callContractFunction(function, contractAddress);
+        } catch (Exception e) {
+            log.warn(e.getMessage());
+        }
+        return token1;
+    }
+
+    private String callContractFunction(Function function, String contractAddress) throws Exception {
+        // encode function
+        String encodedFunction = FunctionEncoder.encode(function);
+        // sign contract
+        Transaction transaction = Transaction.createEthCallTransaction(null, contractAddress, encodedFunction);
+        EthCall response = web3j.ethCall(transaction, DefaultBlockParameterName.LATEST).send();
+        if (response.isReverted())
+            throw new RuntimeException("Error calling contract function: " + response.getRevertReason());
+        // decode return value
+        List<Type> decodedResponse = FunctionReturnDecoder.decode(response.getValue(), function.getOutputParameters());
+        // assume the first return value is address
+        return decodedResponse.get(0).getValue().toString();
     }
 }
